move conversations list to own component, handle case where the list is being reloaded
